+++ kirksdk/qcos_api.go
@@ -302,6 +302,15 @@ type QcosClient interface {
 
 	// GET /v3/alert/containers/<ip>?level=<level>
 	GetContainerAlert(ctx context.Context, ip string, level string) (ret []ContainerAlertInfo, err error)
+
+	// GET /v3/configservices
+	ListConfigServiceSpecs(ctx context.Context) (ret []ConfigServiceSpecInfo, err error)
+
+	// POST /v3/configservices
+	CreateConfigServiceSpec(ctx context.Context, args CreateConfigServiceSpecArgs) (err error)
+
+	// GET /v3/configservices/<namespace>
+	GetConfigServiceSpec(ctx context.Context, namespace string) (ret ConfigServiceSpecInfo, err error)
 }
 
 const (
@@ -407,6 +416,7 @@ type ServiceSpecExport struct {
 	Hosts         []string           `json:"hosts"`
 	Image         string             `json:"image"`
 	LogCollectors []LogCollectorSpec `json:"logCollectors"`
+	Confs         []ConfSpec         `json:"confs,omitempty"`
 	StopGraceSec  int                `json:"stopGraceSec"`
 	WorkDir       string             `json:"workDir"`
 	UnitType      string             `json:"unitType"`
@@ -421,10 +431,16 @@ type ServiceSpec struct {
 	Hosts         []string           `json:"hosts,omitempty"`
 	Image         string             `json:"image,omitempty"`
 	LogCollectors []LogCollectorSpec `json:"logCollectors,omitempty"`
+	Confs         []ConfSpec         `json:"confs,omitempty"`
 	StopGraceSec  int                `json:"stopGraceSec,omitempty"`
 	WorkDir       string             `json:"workDir,omitempty"`
 	UnitType      string             `json:"unitType,omitempty"`
 	GpuUUIDs      []string           `json:"gpuUUIDs,omitempty"` // do not use. prepare for gpu service
+}
+
+type ConfSpec struct {
+	Namespace string   `json:"namespace"`
+	Files     []string `json:"files"`
 }
 
 type LogCollectorSpec struct {
@@ -717,6 +733,7 @@ type JobTaskSpec struct {
 	Envs          []string           `json:"envs,omitempty"`
 	Hosts         []string           `json:"hosts,omitempty"`
 	LogCollectors []LogCollectorSpec `json:"logCollectors,omitempty"`
+	Confs         []ConfSpec         `json:"confs,omitempty"`
 	Deps          []string           `json:"deps,omitempty"`
 	WorkDir       string             `json:"workDir,omitempty"`
 	InstanceNum   int                `json:"instanceNum,omitempty"`
@@ -770,6 +787,7 @@ type JobInstanceID struct {
 type JobTaskSpecEx struct {
 	WorkDir       string             `json:"workDir,omitempty"`
 	LogCollectors []LogCollectorSpec `json:"logCollectors,omitempty"`
+	Confs         []ConfSpec         `json:"confs,omitempty"`
 	Command       []string           `json:"command,omitempty"`
 	EntryPoint    []string           `json:"entryPoint,omitempty"`
 	Envs          []string           `json:"envs,omitempty"`
@@ -860,3 +878,11 @@ type ContainerAlertInfo struct {
 	Threshold *AlertContainerThreshold `json:"threshold"`
 	Methods   []*AlertMethod           `json:"methods"`
 }
+
+type CreateConfigServiceSpecArgs struct {
+	Namespace string                   `json:"namespace"`
+	Vars      map[string]interface{}   `json:"vars"`
+	Listvars  []map[string]interface{} `json:"listvars"`
+}
+
+type ConfigServiceSpecInfo CreateConfigServiceSpecArgs