+++ kirksdk/qcos_client.go
@@ -1252,6 +1252,27 @@ func (p *qcosClientImp) GetContainerAlert(ctx context.Context, ip string, level 
 	return
 }
 
+// GET /v3/configservices
+func (p *qcosClientImp) ListConfigServiceSpecs(ctx context.Context) (ret []ConfigServiceSpecInfo, err error) {
+	url := fmt.Sprintf("%s/v3/configservices")
+	err = p.client.Call(ctx, &ret, "GET", url)
+	return
+}
+
+// POST /v3/configservices
+func (p *qcosClientImp) CreateConfigServiceSpec(ctx context.Context, args CreateConfigServiceSpecArgs) (err error) {
+	url := fmt.Sprintf("%s/v3/configservices")
+	err = p.client.CallWithJson(ctx, nil, "POST", url, args)
+	return
+}
+
+// GET /v3/configservices/<namespace>
+func (p *qcosClientImp) GetConfigServiceSpec(ctx context.Context, namespace string) (ret ConfigServiceSpecInfo, err error) {
+	url := fmt.Sprintf("%s/v3/configservices/%s", namespace)
+	err = p.client.Call(ctx, &ret, "GET", url)
+	return
+}
+
 func (p *qcosClientImp) wait4StackRunning(stackName string, timeout time.Duration) (err error) {
 	if stackName == "" {
 		stackName = DefaultStack