+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # vNext
+- 为 Service 以及 Job 增加 Confs 域
+- 增加 ConfigService 相关 V3 接口
 
 # Release 1.1.0
 - AccountClient 相关 API 使用 appd V3 接口